+++ test/ffprobe.go
@@ -90,120 +90,6 @@ func ffprobe(input string) (*FFProbeInfo, error) {
 	return info, err
 }
 
-func verifyFile(t *testing.T, conf *TestConfig, p *config.PipelineConfig, res *livekit.EgressInfo) {
-	// egress info
-	require.Equal(t, res.Error == "", res.Status != livekit.EgressStatus_EGRESS_FAILED)
-	require.NotZero(t, res.StartedAt)
-	require.NotZero(t, res.EndedAt)
-
-	// file info
-	fileRes := res.GetFile()
-	if fileRes == nil {
-		require.Len(t, res.FileResults, 1)
-		fileRes = res.FileResults[0]
-	}
-
-	require.NotEmpty(t, fileRes.Location)
-	require.Greater(t, fileRes.Size, int64(0))
-	require.Greater(t, fileRes.Duration, int64(0))
-
-	storagePath := fileRes.Filename
-	localPath := fileRes.Filename
-	require.NotEmpty(t, storagePath)
-	require.False(t, strings.Contains(storagePath, "{"))
-
-	// download from cloud storage
-	if uploadConfig := p.GetFileConfig().UploadConfig; uploadConfig != nil {
-		localPath = fmt.Sprintf("%s/%s", conf.LocalOutputDirectory, storagePath)
-		download(t, uploadConfig, localPath, storagePath)
-		download(t, uploadConfig, localPath+".json", storagePath+".json")
-	}
-
-	// verify
-	verify(t, localPath, p, res, types.EgressTypeFile, conf.Muting, conf.sourceFramerate)
-}
-
-func verifyStreams(t *testing.T, p *config.PipelineConfig, conf *TestConfig, urls ...string) {
-	for _, url := range urls {
-		verify(t, url, p, nil, types.EgressTypeStream, false, conf.sourceFramerate)
-	}
-}
-
-func verifySegments(t *testing.T, conf *TestConfig, p *config.PipelineConfig, filenameSuffix livekit.SegmentedFileSuffix, res *livekit.EgressInfo) {
-	// egress info
-	require.Equal(t, res.Error == "", res.Status != livekit.EgressStatus_EGRESS_FAILED)
-	require.NotZero(t, res.StartedAt)
-	require.NotZero(t, res.EndedAt)
-
-	// segments info
-	require.Len(t, res.GetSegmentResults(), 1)
-	segments := res.GetSegmentResults()[0]
-
-	require.NotEmpty(t, segments.PlaylistName)
-	require.NotEmpty(t, segments.PlaylistLocation)
-	require.Greater(t, segments.Size, int64(0))
-	require.Greater(t, segments.Duration, int64(0))
-
-	storedPlaylistPath := segments.PlaylistName
-	localPlaylistPath := segments.PlaylistName
-
-	// download from cloud storage
-	if uploadConfig := p.GetSegmentConfig().UploadConfig; uploadConfig != nil {
-		base := storedPlaylistPath[:len(storedPlaylistPath)-5]
-		localPlaylistPath = fmt.Sprintf("%s/%s", conf.LocalOutputDirectory, storedPlaylistPath)
-		download(t, uploadConfig, localPlaylistPath, storedPlaylistPath)
-		download(t, uploadConfig, localPlaylistPath+".json", storedPlaylistPath+".json")
-		for i := 0; i < int(segments.SegmentCount); i++ {
-			cloudPath := fmt.Sprintf("%s_%05d.ts", base, i)
-			localPath := fmt.Sprintf("%s/%s", conf.LocalOutputDirectory, cloudPath)
-			download(t, uploadConfig, localPath, cloudPath)
-		}
-	}
-
-	verifyPlaylistProgramDateTime(t, filenameSuffix, localPlaylistPath)
-
-	// verify
-	verify(t, localPlaylistPath, p, res, types.EgressTypeSegments, conf.Muting, conf.sourceFramerate)
-}
-
-func verifyPlaylistProgramDateTime(t *testing.T, filenameSuffix livekit.SegmentedFileSuffix, localPlaylistPath string) {
-	p, err := readPlaylist(localPlaylistPath)
-	require.NoError(t, err)
-	require.Equal(t, "EVENT", p.MediaType)
-	require.True(t, p.Closed)
-
-	now := time.Now()
-
-	for i, s := range p.Segments {
-		const leeway = 50 * time.Millisecond
-
-		// Make sure the program date time is current, ie not more than 2 min in the past
-		require.InDelta(t, now.Unix(), s.ProgramDateTime.Unix(), 120)
-
-		if filenameSuffix == livekit.SegmentedFileSuffix_TIMESTAMP {
-			m := segmentTimeRegexp.FindStringSubmatch(s.Filename)
-			require.Equal(t, 3, len(m))
-
-			tm, err := time.Parse("20060102150405", m[1])
-			require.NoError(t, err)
-
-			ms, err := strconv.Atoi(m[2])
-			require.NoError(t, err)
-
-			tm = tm.Add(time.Duration(ms) * time.Millisecond)
-
-			require.InDelta(t, s.ProgramDateTime.UnixNano(), tm.UnixNano(), float64(time.Millisecond))
-		}
-
-		if i < p.Count()-1 {
-			nextSegmentStartDate := p.Segments[i+1].ProgramDateTime
-
-			dateDuration := nextSegmentStartDate.Sub(s.ProgramDateTime)
-			require.InDelta(t, time.Duration(s.Duration*float64(time.Second)), dateDuration, float64(leeway))
-		}
-	}
-}
-
 func verify(t *testing.T, in string, p *config.PipelineConfig, res *livekit.EgressInfo, egressType types.EgressType, withMuting bool, sourceFramerate float64) {
 	info, err := ffprobe(in)
 	require.NoError(t, err, "input %s does not exist", in)
