--- test/helpers.go
@@ -1,297 +1,0 @@
-//go:build integration
-
-package test
-
-import (
-	"context"
-	"encoding/json"
-	"fmt"
-	"io"
-	"net/url"
-	"os"
-	"testing"
-	"time"
-
-	"cloud.google.com/go/storage"
-	"github.com/Azure/azure-storage-blob-go/azblob"
-	"github.com/aws/aws-sdk-go/aws"
-	"github.com/aws/aws-sdk-go/aws/credentials"
-	"github.com/aws/aws-sdk-go/aws/session"
-	"github.com/aws/aws-sdk-go/service/s3"
-	"github.com/aws/aws-sdk-go/service/s3/s3manager"
-	"github.com/googleapis/gax-go/v2"
-	"github.com/stretchr/testify/require"
-	"google.golang.org/api/option"
-
-	"github.com/livekit/egress/pkg/config"
-	"github.com/livekit/egress/pkg/service"
-	"github.com/livekit/egress/pkg/types"
-	"github.com/livekit/protocol/livekit"
-	"github.com/livekit/protocol/logger"
-	"github.com/livekit/protocol/rpc"
-	lksdk "github.com/livekit/server-sdk-go"
-)
-
-const muteDuration = time.Second * 10
-
-var (
-	samples = map[types.MimeType]string{
-		types.MimeTypeOpus: "/workspace/test/sample/matrix-trailer.ogg",
-		types.MimeTypeH264: "/workspace/test/sample/matrix-trailer.h264",
-		types.MimeTypeVP8:  "/workspace/test/sample/matrix-trailer.ivf",
-	}
-
-	frameDurations = map[types.MimeType]time.Duration{
-		types.MimeTypeH264: time.Microsecond * 41708,
-		types.MimeTypeVP8:  time.Microsecond * 41708,
-	}
-)
-
-func publishSamplesToRoom(t *testing.T, room *lksdk.Room, audioCodec, videoCodec types.MimeType, withMuting bool) (audioTrackID, videoTrackID string) {
-	withVideoMuting := false
-	if audioCodec == "" {
-		withVideoMuting = withMuting
-	}
-	if videoCodec != "" {
-		videoTrackID = publishSampleToRoom(t, room, videoCodec, withVideoMuting)
-	}
-	if audioCodec != "" {
-		audioTrackID = publishSampleToRoom(t, room, audioCodec, withMuting)
-	}
-
-	time.Sleep(time.Second)
-	return
-}
-
-func publishSampleToRoom(t *testing.T, room *lksdk.Room, codec types.MimeType, withMuting bool) string {
-	filename := samples[codec]
-	frameDuration := frameDurations[codec]
-
-	var pub *lksdk.LocalTrackPublication
-	done := make(chan struct{})
-	opts := []lksdk.ReaderSampleProviderOption{
-		lksdk.ReaderTrackWithOnWriteComplete(func() {
-			close(done)
-			if pub != nil {
-				_ = room.LocalParticipant.UnpublishTrack(pub.SID())
-			}
-		}),
-	}
-
-	if frameDuration != 0 {
-		opts = append(opts, lksdk.ReaderTrackWithFrameDuration(frameDuration))
-	}
-
-	track, err := lksdk.NewLocalFileTrack(filename, opts...)
-	require.NoError(t, err)
-
-	pub, err = room.LocalParticipant.PublishTrack(track, &lksdk.TrackPublicationOptions{Name: filename})
-	require.NoError(t, err)
-
-	trackID := pub.SID()
-	t.Cleanup(func() {
-		_ = room.LocalParticipant.UnpublishTrack(trackID)
-	})
-
-	if withMuting {
-		go func() {
-			muted := false
-			time.Sleep(time.Second * 15)
-			for {
-				select {
-				case <-done:
-					return
-				default:
-					pub.SetMuted(!muted)
-					muted = !muted
-					time.Sleep(muteDuration)
-				}
-			}
-		}()
-	}
-
-	return trackID
-}
-
-func getFilePath(conf *config.ServiceConfig, filename string) string {
-	if conf.S3 != nil || conf.Azure != nil || conf.GCP != nil || conf.AliOSS != nil {
-		return filename
-	}
-
-	return fmt.Sprintf("%s/%s", conf.LocalOutputDirectory, filename)
-}
-
-func startEgress(t *testing.T, conf *TestConfig, req *rpc.StartEgressRequest) string {
-	// send start request
-	info, err := conf.client.StartEgress(context.Background(), "", req)
-
-	// check returned egress info
-	require.NoError(t, err)
-	require.Empty(t, info.Error)
-	require.NotEmpty(t, info.EgressId)
-	switch req.Request.(type) {
-	case *rpc.StartEgressRequest_Web:
-		require.Empty(t, info.RoomName)
-	default:
-		require.Equal(t, conf.RoomName, info.RoomName)
-	}
-
-	require.Equal(t, livekit.EgressStatus_EGRESS_STARTING.String(), info.Status.String())
-
-	// check status
-	if conf.HealthPort != 0 {
-		status := getStatus(t, conf.svc)
-		require.Contains(t, status, info.EgressId)
-	}
-
-	// wait
-	time.Sleep(time.Second * 5)
-
-	// check active update
-	checkUpdate(t, conf, info.EgressId, livekit.EgressStatus_EGRESS_ACTIVE)
-
-	return info.EgressId
-}
-
-func getStatus(t *testing.T, svc *service.Service) map[string]interface{} {
-	b, err := svc.Status()
-	require.NoError(t, err)
-
-	status := make(map[string]interface{})
-	err = json.Unmarshal(b, &status)
-	require.NoError(t, err)
-
-	return status
-}
-
-func stopEgress(t *testing.T, conf *TestConfig, egressID string) *livekit.EgressInfo {
-	// send stop request
-	info, err := conf.client.StopEgress(context.Background(), egressID, &livekit.StopEgressRequest{
-		EgressId: egressID,
-	})
-
-	// check returned egress info
-	require.NoError(t, err)
-	require.Empty(t, info.Error)
-	require.NotEmpty(t, info.StartedAt)
-	require.Equal(t, livekit.EgressStatus_EGRESS_ENDING.String(), info.Status.String())
-
-	// check complete update
-	return checkStoppedEgress(t, conf, egressID, livekit.EgressStatus_EGRESS_COMPLETE)
-}
-
-func download(t *testing.T, uploadParams interface{}, localFilepath, storageFilepath string) {
-	logger.Debugw("download", "localFilepath", localFilepath, "storageFilepath", storageFilepath)
-	switch u := uploadParams.(type) {
-	case *livekit.S3Upload:
-		downloadS3(t, u, localFilepath, storageFilepath)
-
-	case *livekit.GCPUpload:
-		downloadGCP(t, u, localFilepath, storageFilepath)
-
-	case *livekit.AzureBlobUpload:
-		downloadAzure(t, u, localFilepath, storageFilepath)
-	}
-}
-
-func downloadS3(t *testing.T, conf *livekit.S3Upload, localFilepath, storageFilepath string) {
-	sess, err := session.NewSession(&aws.Config{
-		Credentials: credentials.NewStaticCredentials(conf.AccessKey, conf.Secret, ""),
-		Endpoint:    aws.String(conf.Endpoint),
-		Region:      aws.String(conf.Region),
-		MaxRetries:  aws.Int(maxRetries),
-	})
-	require.NoError(t, err)
-
-	file, err := os.Create(localFilepath)
-	require.NoError(t, err)
-	defer file.Close()
-
-	_, err = s3manager.NewDownloader(sess).Download(file,
-		&s3.GetObjectInput{
-			Bucket: aws.String(conf.Bucket),
-			Key:    aws.String(storageFilepath),
-		},
-	)
-	require.NoError(t, err)
-
-	_, err = s3.New(sess).DeleteObject(&s3.DeleteObjectInput{
-		Bucket: aws.String(conf.Bucket),
-		Key:    aws.String(storageFilepath),
-	})
-	require.NoError(t, err)
-}
-
-func downloadAzure(t *testing.T, conf *livekit.AzureBlobUpload, localFilepath, storageFilepath string) {
-	credential, err := azblob.NewSharedKeyCredential(
-		conf.AccountName,
-		conf.AccountKey,
-	)
-	require.NoError(t, err)
-
-	pipeline := azblob.NewPipeline(credential, azblob.PipelineOptions{
-		Retry: azblob.RetryOptions{
-			Policy:        azblob.RetryPolicyExponential,
-			MaxTries:      maxRetries,
-			MaxRetryDelay: maxDelay,
-		},
-	})
-	sUrl := fmt.Sprintf("https://%s.blob.core.windows.net/%s", conf.AccountName, conf.ContainerName)
-	azUrl, err := url.Parse(sUrl)
-	require.NoError(t, err)
-
-	containerURL := azblob.NewContainerURL(*azUrl, pipeline)
-	blobURL := containerURL.NewBlobURL(storageFilepath)
-
-	file, err := os.Create(localFilepath)
-	require.NoError(t, err)
-	defer file.Close()
-
-	err = azblob.DownloadBlobToFile(context.Background(), blobURL, 0, 0, file, azblob.DownloadFromBlobOptions{
-		BlockSize:   4 * 1024 * 1024,
-		Parallelism: 16,
-		RetryReaderOptionsPerBlock: azblob.RetryReaderOptions{
-			MaxRetryRequests: 3,
-		},
-	})
-	require.NoError(t, err)
-
-	_, err = blobURL.Delete(context.Background(), azblob.DeleteSnapshotsOptionNone, azblob.BlobAccessConditions{})
-	require.NoError(t, err)
-}
-
-func downloadGCP(t *testing.T, conf *livekit.GCPUpload, localFilepath, storageFilepath string) {
-	ctx := context.Background()
-	var client *storage.Client
-
-	var err error
-	if conf.Credentials != "" {
-		client, err = storage.NewClient(ctx, option.WithCredentialsJSON([]byte(conf.Credentials)))
-	} else {
-		client, err = storage.NewClient(ctx)
-	}
-	require.NoError(t, err)
-	defer client.Close()
-
-	file, err := os.Create(localFilepath)
-	require.NoError(t, err)
-	defer file.Close()
-
-	rc, err := client.Bucket(conf.Bucket).Object(storageFilepath).Retryer(
-		storage.WithBackoff(
-			gax.Backoff{
-				Initial:    minDelay,
-				Max:        maxDelay,
-				Multiplier: 2,
-			}),
-		storage.WithPolicy(storage.RetryAlways),
-	).NewReader(ctx)
-	require.NoError(t, err)
-
-	_, err = io.Copy(file, rc)
-	_ = rc.Close()
-	require.NoError(t, err)
-
-	err = client.Bucket(conf.Bucket).Object(storageFilepath).Delete(context.Background())
-	require.NoError(t, err)
-}
