+++ test/integration.go
@@ -4,34 +4,45 @@ package test
 
 import (
 	"context"
+	"encoding/json"
 	"fmt"
-	"io/fs"
-	"math/rand"
 	"testing"
 	"time"
 
-	"github.com/go-logr/logr"
 	"github.com/stretchr/testify/require"
 
 	"github.com/livekit/egress/pkg/config"
-	"github.com/livekit/egress/pkg/service"
 	"github.com/livekit/egress/pkg/types"
 	"github.com/livekit/protocol/livekit"
-	"github.com/livekit/protocol/logger"
 	"github.com/livekit/protocol/rpc"
-	"github.com/livekit/protocol/utils"
-	"github.com/livekit/psrpc"
 	lksdk "github.com/livekit/server-sdk-go"
 )
 
 const (
-	streamUrl1     = "rtmp://localhost:1935/live/stream1"
-	redactedUrl1   = "rtmp://localhost:1935/live/*******"
-	streamUrl2     = "rtmp://localhost:1935/live/stream10"
-	redactedUrl2   = "rtmp://localhost:1935/live/********"
-	badStreamUrl   = "rtmp://sfo.contribute.live-video.net/app/fake1"
-	redactedBadUrl = "rtmp://sfo.contribute.live-video.net/app/*****"
-	webUrl         = "https://www.youtube.com/watch?v=wjQq0nSGS28&t=5205s"
+	muteDuration = time.Second * 10
+
+	streamUrl1      = "rtmp://localhost:1935/live/stream1"
+	redactedUrl1    = "rtmp://localhost:1935/live/*******"
+	streamUrl2      = "rtmp://localhost:1935/live/stream10"
+	redactedUrl2    = "rtmp://localhost:1935/live/********"
+	badStreamUrl1   = "rtmp://sfo.contribute.live-video.net/app/fake1"
+	redactedBadUrl1 = "rtmp://sfo.contribute.live-video.net/app/*****"
+	badStreamUrl2   = "rtmp://localhost:1936/live/stream"
+	redactedBadUrl2 = "rtmp://localhost:1936/live/******"
+	webUrl          = "https://www.youtube.com/watch?v=wjQq0nSGS28&t=5205s"
+)
+
+var (
+	samples = map[types.MimeType]string{
+		types.MimeTypeOpus: "/workspace/test/sample/matrix-trailer.ogg",
+		types.MimeTypeH264: "/workspace/test/sample/matrix-trailer.h264",
+		types.MimeTypeVP8:  "/workspace/test/sample/matrix-trailer.ivf",
+	}
+
+	frameDurations = map[types.MimeType]time.Duration{
+		types.MimeTypeH264: time.Microsecond * 41708,
+		types.MimeTypeVP8:  time.Microsecond * 41708,
+	}
 )
 
 type testCase struct {
@@ -60,90 +71,28 @@ type testCase struct {
 	expectVideoTranscoding bool
 }
 
-func RunTestSuite(t *testing.T, conf *TestConfig, bus psrpc.MessageBus, templateFs fs.FS) {
-	lksdk.SetLogger(logger.LogRLogger(logr.Discard()))
-
-	// connect to room
-	room, err := lksdk.ConnectToRoom(conf.WsUrl, lksdk.ConnectInfo{
-		APIKey:              conf.ApiKey,
-		APISecret:           conf.ApiSecret,
-		RoomName:            conf.RoomName,
-		ParticipantName:     "egress-sample",
-		ParticipantIdentity: fmt.Sprintf("sample-%d", rand.Intn(100)),
-	}, lksdk.NewRoomCallback())
-	require.NoError(t, err)
-	defer room.Disconnect()
-
-	// start service
-	ioClient, err := rpc.NewIOInfoClient("test_io_client", bus)
-	require.NoError(t, err)
-	svc, err := service.NewService(conf.ServiceConfig, bus, nil, ioClient)
-	require.NoError(t, err)
-
-	psrpcClient, err := rpc.NewEgressClient(livekit.NodeID(utils.NewGuid("TEST_")), bus)
-	require.NoError(t, err)
-
-	// start debug handler
-	svc.StartDebugHandlers()
-
-	// start templates handler
-	err = svc.StartTemplatesServer(templateFs)
-	require.NoError(t, err)
-
-	go func() {
-		err := svc.Run()
-		require.NoError(t, err)
-	}()
-	t.Cleanup(func() { svc.Stop(true) })
-	time.Sleep(time.Second * 3)
-
-	// subscribe to update channel
-	psrpcUpdates := make(chan *livekit.EgressInfo, 100)
-	_, err = newIOTestServer(bus, psrpcUpdates)
-	require.NoError(t, err)
-
-	// update test config
-	conf.svc = svc
-	conf.client = psrpcClient
-	conf.updates = psrpcUpdates
-	conf.room = room
-
-	// check status
-	if conf.HealthPort != 0 {
-		status := getStatus(t, svc)
-		require.Len(t, status, 1)
-		require.Contains(t, status, "CpuLoad")
-	}
-
-	// run tests
-	runRoomCompositeTests(t, conf)
-	runWebTests(t, conf)
-	runTrackCompositeTests(t, conf)
-	runTrackTests(t, conf)
-}
-
-func runWebTest(t *testing.T, conf *TestConfig, name string, audioCodec, videoCodec types.MimeType, f func(t *testing.T)) {
+func (r *Runner) runWebTest(t *testing.T, name string, audioCodec, videoCodec types.MimeType, f func(t *testing.T)) {
 	t.Run(name, func(t *testing.T) {
-		awaitIdle(t, conf.svc)
-		publishSamplesToRoom(t, conf.room, audioCodec, videoCodec, conf.Muting)
+		r.awaitIdle(t)
+		r.publishSamplesToRoom(t, audioCodec, videoCodec)
 		f(t)
 	})
 }
 
-func runSDKTest(t *testing.T, conf *TestConfig, name string, audioCodec, videoCodec types.MimeType,
+func (r *Runner) runSDKTest(t *testing.T, name string, audioCodec, videoCodec types.MimeType,
 	f func(t *testing.T, audioTrackID, videoTrackID string),
 ) {
 	t.Run(name, func(t *testing.T) {
-		awaitIdle(t, conf.svc)
-		audioTrackID, videoTrackID := publishSamplesToRoom(t, conf.room, audioCodec, videoCodec, conf.Muting)
+		r.awaitIdle(t)
+		audioTrackID, videoTrackID := r.publishSamplesToRoom(t, audioCodec, videoCodec)
 		f(t, audioTrackID, videoTrackID)
 	})
 }
 
-func awaitIdle(t *testing.T, svc *service.Service) {
-	svc.KillAll()
+func (r *Runner) awaitIdle(t *testing.T) {
+	r.svc.KillAll()
 	for i := 0; i < 30; i++ {
-		status := getStatus(t, svc)
+		status := r.getStatus(t)
 		if len(status) == 1 {
 			return
 		}
@@ -152,215 +101,210 @@ func awaitIdle(t *testing.T, svc *service.Service) {
 	t.Fatal("service not idle after 30s")
 }
 
-func runFileTest(t *testing.T, conf *TestConfig, req *rpc.StartEgressRequest, test *testCase) {
-	conf.SessionLimits.FileOutputMaxDuration = test.sessionTimeout
-
-	// start
-	egressID := startEgress(t, conf, req)
-
-	var res *livekit.EgressInfo
-	if conf.SessionLimits.FileOutputMaxDuration > 0 {
-		time.Sleep(conf.SessionLimits.FileOutputMaxDuration + time.Second)
-
-		res = checkStoppedEgress(t, conf, egressID, livekit.EgressStatus_EGRESS_LIMIT_REACHED)
-	} else {
-		time.Sleep(time.Second * 25)
-
-		// stop
-		res = stopEgress(t, conf, egressID)
+func (r *Runner) publishSamplesToRoom(t *testing.T, audioCodec, videoCodec types.MimeType) (audioTrackID, videoTrackID string) {
+	withVideoMuting := false
+	if audioCodec == "" {
+		withVideoMuting = r.Muting
+	}
+	if videoCodec != "" {
+		videoTrackID = r.publishSampleToRoom(t, videoCodec, withVideoMuting)
+	}
+	if audioCodec != "" {
+		audioTrackID = r.publishSampleToRoom(t, audioCodec, r.Muting)
 	}
 
-	// get params
-	p, err := config.GetValidatedPipelineConfig(conf.ServiceConfig, req)
-	require.NoError(t, err)
-	if p.GetFileConfig().OutputType == types.OutputTypeUnknownFile {
-		p.GetFileConfig().OutputType = test.outputType
+	time.Sleep(time.Second)
+	return
+}
+
+func (r *Runner) publishSampleToRoom(t *testing.T, codec types.MimeType, withMuting bool) string {
+	filename := samples[codec]
+	frameDuration := frameDurations[codec]
+
+	var pub *lksdk.LocalTrackPublication
+	done := make(chan struct{})
+	opts := []lksdk.ReaderSampleProviderOption{
+		lksdk.ReaderTrackWithOnWriteComplete(func() {
+			close(done)
+			if pub != nil {
+				_ = r.room.LocalParticipant.UnpublishTrack(pub.SID())
+			}
+		}),
 	}
 
-	require.Equal(t, test.expectVideoTranscoding, p.VideoTranscoding)
-
-	// verify
-	verifyFile(t, conf, p, res)
-}
-
-func runStreamTest(t *testing.T, conf *TestConfig, req *rpc.StartEgressRequest, test *testCase) {
-	conf.SessionLimits.StreamOutputMaxDuration = test.sessionTimeout
-
-	if conf.SessionLimits.StreamOutputMaxDuration > 0 {
-		runTimeLimitStreamTest(t, conf, req, test)
-	} else {
-		runMultipleStreamTest(t, conf, req, test)
+	if frameDuration != 0 {
+		opts = append(opts, lksdk.ReaderTrackWithFrameDuration(frameDuration))
 	}
-}
 
-func runTimeLimitStreamTest(t *testing.T, conf *TestConfig, req *rpc.StartEgressRequest, test *testCase) {
-	egressID := startEgress(t, conf, req)
-
-	time.Sleep(time.Second * 5)
-
-	// get params
-	p, err := config.GetValidatedPipelineConfig(conf.ServiceConfig, req)
+	track, err := lksdk.NewLocalFileTrack(filename, opts...)
 	require.NoError(t, err)
 
-	require.Equal(t, test.expectVideoTranscoding, p.VideoTranscoding)
-
-	verifyStreams(t, p, conf, streamUrl1)
-
-	time.Sleep(conf.SessionLimits.StreamOutputMaxDuration - time.Second*4)
-
-	checkStoppedEgress(t, conf, egressID, livekit.EgressStatus_EGRESS_LIMIT_REACHED)
-}
-
-func runMultipleStreamTest(t *testing.T, conf *TestConfig, req *rpc.StartEgressRequest, test *testCase) {
-	ctx := context.Background()
-	egressID := startEgress(t, conf, req)
-
-	time.Sleep(time.Second * 5)
-
-	// get params
-	p, err := config.GetValidatedPipelineConfig(conf.ServiceConfig, req)
+	pub, err = r.room.LocalParticipant.PublishTrack(track, &lksdk.TrackPublicationOptions{Name: filename})
 	require.NoError(t, err)
 
-	// verify stream
-	require.Equal(t, test.expectVideoTranscoding, p.VideoTranscoding)
-	verifyStreams(t, p, conf, streamUrl1)
-
-	// add one good stream url and a couple bad ones
-	_, err = conf.client.UpdateStream(ctx, egressID, &livekit.UpdateStreamRequest{
-		EgressId:      egressID,
-		AddOutputUrls: []string{badStreamUrl, streamUrl2},
+	trackID := pub.SID()
+	t.Cleanup(func() {
+		_ = r.room.LocalParticipant.UnpublishTrack(trackID)
 	})
-	require.NoError(t, err)
 
-	time.Sleep(time.Second * 5)
-
-	update := getUpdate(t, conf, egressID)
-	require.Equal(t, livekit.EgressStatus_EGRESS_ACTIVE.String(), update.Status.String())
-	require.Len(t, update.StreamResults, 3)
-	for _, info := range update.StreamResults {
-		switch info.Url {
-		case redactedUrl1, redactedUrl2:
-			require.Equal(t, livekit.StreamInfo_ACTIVE.String(), info.Status.String())
-
-		case redactedBadUrl:
-			require.Equal(t, livekit.StreamInfo_FAILED.String(), info.Status.String())
-
-		default:
-			t.Fatal("invalid stream url in result")
-		}
+	if withMuting {
+		go func() {
+			muted := false
+			time.Sleep(time.Second * 15)
+			for {
+				select {
+				case <-done:
+					return
+				default:
+					pub.SetMuted(!muted)
+					muted = !muted
+					time.Sleep(muteDuration)
+				}
+			}
+		}()
 	}
 
-	require.Equal(t, test.expectVideoTranscoding, p.VideoTranscoding)
-
-	// verify the good stream urls
-	verifyStreams(t, p, conf, streamUrl1, streamUrl2)
-
-	// remove one of the stream urls
-	_, err = conf.client.UpdateStream(ctx, egressID, &livekit.UpdateStreamRequest{
-		EgressId:         egressID,
-		RemoveOutputUrls: []string{streamUrl1},
-	})
-	require.NoError(t, err)
-
-	time.Sleep(time.Second * 5)
-
-	// verify the remaining stream
-	verifyStreams(t, p, conf, streamUrl2)
-
-	time.Sleep(time.Second * 10)
-
-	// stop
-	res := stopEgress(t, conf, egressID)
-
-	// verify egress info
-	require.Empty(t, res.Error)
-	require.NotZero(t, res.StartedAt)
-	require.NotZero(t, res.EndedAt)
-
-	// check stream info
-	require.Len(t, res.StreamResults, 3)
-	for _, info := range res.StreamResults {
-		require.NotZero(t, info.StartedAt)
-		require.NotZero(t, info.EndedAt)
-
-		switch info.Url {
-		case redactedUrl1:
-			require.Equal(t, livekit.StreamInfo_FINISHED.String(), info.Status.String())
-			require.Greater(t, float64(info.Duration)/1e9, 15.0)
-
-		case redactedUrl2:
-			require.Equal(t, livekit.StreamInfo_FINISHED.String(), info.Status.String())
-			require.Greater(t, float64(info.Duration)/1e9, 10.0)
-
-		case redactedBadUrl:
-			require.Equal(t, livekit.StreamInfo_FAILED.String(), info.Status.String())
-
-		default:
-			t.Fatal("invalid stream url in result")
-		}
-	}
+	return trackID
 }
 
-func runSegmentsTest(t *testing.T, conf *TestConfig, req *rpc.StartEgressRequest, test *testCase) {
-	conf.SessionLimits.SegmentOutputMaxDuration = test.sessionTimeout
-
-	egressID := startEgress(t, conf, req)
-
-	var res *livekit.EgressInfo
-	if conf.SessionLimits.SegmentOutputMaxDuration > 0 {
-		time.Sleep(conf.SessionLimits.SegmentOutputMaxDuration + time.Second)
-
-		res = checkStoppedEgress(t, conf, egressID, livekit.EgressStatus_EGRESS_LIMIT_REACHED)
-	} else {
-		time.Sleep(time.Second * 25)
-
-		// stop
-		res = stopEgress(t, conf, egressID)
-	}
-
-	// get params
-	p, err := config.GetValidatedPipelineConfig(conf.ServiceConfig, req)
-	require.NoError(t, err)
-
-	require.Equal(t, test.expectVideoTranscoding, p.VideoTranscoding)
-	verifySegments(t, conf, p, test.filenameSuffix, res)
-}
-
-func runMultipleTest(
+func (r *Runner) runMultipleTest(
 	t *testing.T,
-	conf *TestConfig,
 	req *rpc.StartEgressRequest,
 	file, stream, segments bool,
 	filenameSuffix livekit.SegmentedFileSuffix,
 ) {
-	conf.SessionLimits = config.SessionLimits{}
-
-	egressID := startEgress(t, conf, req)
+	egressID := r.startEgress(t, req)
 	time.Sleep(time.Second * 10)
 
 	// get params
-	p, err := config.GetValidatedPipelineConfig(conf.ServiceConfig, req)
+	p, err := config.GetValidatedPipelineConfig(r.ServiceConfig, req)
 	require.NoError(t, err)
 
 	if stream {
-		_, err := conf.client.UpdateStream(context.Background(), egressID, &livekit.UpdateStreamRequest{
+		_, err := r.client.UpdateStream(context.Background(), egressID, &livekit.UpdateStreamRequest{
 			EgressId:      egressID,
 			AddOutputUrls: []string{streamUrl1},
 		})
 		require.NoError(t, err)
 
 		time.Sleep(time.Second * 10)
-		verifyStreams(t, p, conf, streamUrl1)
+		r.verifyStreams(t, p, streamUrl1)
 		time.Sleep(time.Second * 10)
 	} else {
 		time.Sleep(time.Second * 20)
 	}
 
-	res := stopEgress(t, conf, egressID)
+	res := r.stopEgress(t, egressID)
 	if file {
-		verifyFile(t, conf, p, res)
+		r.verifyFile(t, p, res)
 	}
 	if segments {
-		verifySegments(t, conf, p, filenameSuffix, res)
+		r.verifySegments(t, p, filenameSuffix, res)
 	}
+}
+
+func (r *Runner) startEgress(t *testing.T, req *rpc.StartEgressRequest) string {
+	// send start request
+	info, err := r.client.StartEgress(context.Background(), "", req)
+
+	// check returned egress info
+	require.NoError(t, err)
+	require.Empty(t, info.Error)
+	require.NotEmpty(t, info.EgressId)
+	switch req.Request.(type) {
+	case *rpc.StartEgressRequest_Web:
+		require.Empty(t, info.RoomName)
+	default:
+		require.Equal(t, r.RoomName, info.RoomName)
+	}
+
+	require.Equal(t, livekit.EgressStatus_EGRESS_STARTING.String(), info.Status.String())
+
+	// check status
+	if r.HealthPort != 0 {
+		status := r.getStatus(t)
+		require.Contains(t, status, info.EgressId)
+	}
+
+	// wait
+	time.Sleep(time.Second * 5)
+
+	// check active update
+	r.checkUpdate(t, info.EgressId, livekit.EgressStatus_EGRESS_ACTIVE)
+
+	return info.EgressId
+}
+
+func (r *Runner) checkUpdate(t *testing.T, egressID string, status livekit.EgressStatus) *livekit.EgressInfo {
+	info := r.getUpdate(t, egressID)
+
+	require.Equal(t, status.String(), info.Status.String())
+	if info.Status == livekit.EgressStatus_EGRESS_FAILED {
+		require.NotEmpty(t, info.Error, "failed egress missing error")
+	} else {
+		require.Empty(t, info.Error, "status %s with error %s", info.Status.String(), info.Error)
+	}
+
+	return info
+}
+
+func (r *Runner) getUpdate(t *testing.T, egressID string) *livekit.EgressInfo {
+	for {
+		select {
+		case info := <-r.updates:
+			if info.EgressId == egressID {
+				return info
+			}
+
+		case <-time.After(time.Minute):
+			t.Fatal("no update from results channel")
+			return nil
+		}
+	}
+}
+
+func (r *Runner) getStatus(t *testing.T) map[string]interface{} {
+	b, err := r.svc.Status()
+	require.NoError(t, err)
+
+	status := make(map[string]interface{})
+	err = json.Unmarshal(b, &status)
+	require.NoError(t, err)
+
+	return status
+}
+
+func (r *Runner) stopEgress(t *testing.T, egressID string) *livekit.EgressInfo {
+	// send stop request
+	info, err := r.client.StopEgress(context.Background(), egressID, &livekit.StopEgressRequest{
+		EgressId: egressID,
+	})
+
+	// check returned egress info
+	require.NoError(t, err)
+	require.Empty(t, info.Error)
+	require.NotEmpty(t, info.StartedAt)
+	require.Equal(t, livekit.EgressStatus_EGRESS_ENDING.String(), info.Status.String())
+
+	// check ending update
+	r.checkUpdate(t, egressID, livekit.EgressStatus_EGRESS_ENDING)
+
+	// get final info
+	res := r.checkUpdate(t, egressID, livekit.EgressStatus_EGRESS_COMPLETE)
+
+	// check status
+	if r.HealthPort != 0 {
+		status := r.getStatus(t)
+		require.Len(t, status, 1)
+	}
+
+	return res
+}
+
+func getFilePath(conf *config.ServiceConfig, filename string) string {
+	if conf.S3 != nil || conf.Azure != nil || conf.GCP != nil || conf.AliOSS != nil {
+		return filename
+	}
+
+	return fmt.Sprintf("%s/%s", conf.LocalOutputDirectory, filename)
 }
